requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
scipy
scikit-fuzzy
torch --index-url https://download.pytorch.org/whl/cpu
jinja2